+++ src/core/bind_management_server.cpp
@@ -4,7 +4,7 @@
  *
  * This file contains the implementation of the Bind Management : Server Role.
  *
- * @version    1.1.1
+ * @version    1.2.0
  *
  * @copyright  Copyright &copy; &nbsp; 2014 ULE Alliance
  *
@@ -50,7 +50,7 @@ HF::Devices::Concentrator::IUnit0 &IServer::unit0 () const
  */
 // =============================================================================
 Common::Result AbstractServer::handle_command (Protocol::Packet &packet,
-                                               Common::ByteArray &payload, size_t offset)
+                                               Common::ByteArray &payload, uint16_t offset)
 {
    Common::Result res      = Common::Result::FAIL_UNKNOWN;
